check deposit: allow additional vouts in coinbase
For some reason, we were not allowed to have additional
vouts in the coinbase transaction if their total was
greater than 0.1 KMD. Therefore, a coinbase transaction
with 2.8 KMD in vout[0] and 0.2 KMD in vout[1] was forbidden.

Since this release, we are allowing such transactions.
We are not using height activation here to keep the code simple;
instead, we have just removed the restriction itself.