enable KIP0003 consensus rule
The KIP0003 consensus rule was added in the `dev` branch
as "logging only", i.e. as disabled. However, in the
`patch-s8-prepare-hf-test` branch, which we used for
testing on hfnet, it's already enabled
(as seen in [this commit](https://github.com/KomodoPlatform/komodo/commit/783fc1f83144d05a1dc6103d5dae4ef677bf241f)).
For the release, we should enable this rule and make the
logging optional using the hfnet category.

- https://github.com/KomodoPlatform/komodo/pull/626