Remove signedmasks file creation
As it turns out, `signedmasks` is no longer used. As a result, it is written in the daemon, but no other component uses it. Additionally, the original implementation left file descriptors open throughout the application's lifecycle and did not close them on exit. Therefore, we have decided to remove this feature from the codebase.